Skiller ut mottak av inntektsmelding til egen InntektsmeldingMottakTjeneste
