+++ src/test/java/no/nav/familie/inntektsmelding/imdialog/tjenester/InntektsmeldingMottakTjenesteTest.java
@@ -1,0 +1,101 @@
+package no.nav.familie.inntektsmelding.imdialog.tjenester;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.when;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import no.nav.familie.inntektsmelding.forespørsel.modell.ForespørselMapper;
+import no.nav.familie.inntektsmelding.forespørsel.tjenester.ForespørselBehandlingTjeneste;
+import no.nav.familie.inntektsmelding.imdialog.modell.InntektsmeldingRepository;
+import no.nav.familie.inntektsmelding.imdialog.rest.SendInntektsmeldingRequestDto;
+import no.nav.familie.inntektsmelding.koder.ForespørselStatus;
+import no.nav.familie.inntektsmelding.koder.Ytelsetype;
+import no.nav.familie.inntektsmelding.typer.dto.AktørIdDto;
+import no.nav.familie.inntektsmelding.typer.dto.ArbeidsgiverDto;
+import no.nav.familie.inntektsmelding.typer.dto.YtelseTypeDto;
+import no.nav.vedtak.felles.prosesstask.api.ProsessTaskTjeneste;
+import no.nav.vedtak.sikkerhet.kontekst.IdentType;
+import no.nav.vedtak.sikkerhet.kontekst.KontekstHolder;
+import no.nav.vedtak.sikkerhet.kontekst.RequestKontekst;
+import no.nav.vedtak.sikkerhet.oidc.config.OpenIDProvider;
+import no.nav.vedtak.sikkerhet.oidc.token.OpenIDToken;
+import no.nav.vedtak.sikkerhet.oidc.token.TokenString;
+
+@ExtendWith(MockitoExtension.class)
+class InntektsmeldingMottakTjenesteTest {
+
+    private static final String INNMELDER_UID = "12324312345";
+
+    @Mock
+    private ForespørselBehandlingTjeneste forespørselBehandlingTjeneste;
+    @Mock
+    private InntektsmeldingRepository inntektsmeldingRepository;
+    @Mock
+    private ProsessTaskTjeneste prosessTaskTjeneste;
+
+    private InntektsmeldingMottakTjeneste inntektsmeldingMottakTjeneste;
+
+    @BeforeAll
+    static void beforeAll() {
+        KontekstHolder.setKontekst(RequestKontekst.forRequest(INNMELDER_UID, "kompakt", IdentType.EksternBruker,
+            new OpenIDToken(OpenIDProvider.TOKENX, new TokenString("token")), UUID.randomUUID(), Set.of()));
+    }
+
+    @AfterAll
+    static void afterAll() {
+        KontekstHolder.fjernKontekst();
+    }
+
+    @BeforeEach
+    void setUp() {
+        inntektsmeldingMottakTjeneste = new InntektsmeldingMottakTjeneste(forespørselBehandlingTjeneste, inntektsmeldingRepository, prosessTaskTjeneste);
+    }
+
+
+    @Test
+    void skal_ikke_godta_im_på_utgått_forespørrsel() {
+        // Arrange
+        var uuid = UUID.randomUUID();
+        var forespørsel = ForespørselMapper.mapForespørsel("999999999",
+            LocalDate.now(),
+            "9999999999999",
+            Ytelsetype.PLEIEPENGER_SYKT_BARN,
+            "123",
+            null,
+            null);
+        forespørsel.setStatus(ForespørselStatus.UTGÅTT);
+        when(forespørselBehandlingTjeneste.hentForespørsel(uuid)).thenReturn(Optional.of(forespørsel));
+        var innsendingDto = new SendInntektsmeldingRequestDto(uuid,
+            new AktørIdDto("9999999999999"),
+            YtelseTypeDto.PLEIEPENGER_SYKT_BARN,
+            new ArbeidsgiverDto("999999999"),
+            new SendInntektsmeldingRequestDto.KontaktpersonRequestDto("Navn", "123"),
+            LocalDate.now(),
+            BigDecimal.valueOf(10000),
+            List.of(),
+            List.of(),
+            List.of(),
+            null);
+
+        // Act
+        var ex = assertThrows(IllegalStateException.class, () -> inntektsmeldingMottakTjeneste.mottaInntektsmelding(innsendingDto));
+
+        // Assert
+        assertThat(ex.getMessage()).contains("Kan ikke motta nye inntektsmeldinger på utgåtte forespørsler");
+    }
+}
